+++ apps/vercel/contentful-app-manifest.json
@@ -1,13 +1,13 @@
 {
   "actions": [
     {
-      "id": "vercelGetPreviewEnvironment",
-      "name": "Get signed URL tokens",
+      "id": "vercelGetPreviewEnvironments",
+      "name": "Get preview environments provided by the Vercel app",
       "category": "Custom",
-      "description": "This action returns signed JWT tokens for securing Mux video playback",
+      "description": "This action returns preview environments provided by Vercel",
       "type": "function",
-      "path": "actions/get-preview-env.js",
-      "entryFile": "./app-actions/src/actions/get-preview-env.ts",
+      "path": "actions/get-preview-envs.js",
+      "entryFile": "./app-actions/src/actions/get-preview-envs.ts",
       "allowNetworks": [],
       "parameters": [ ]
     }
